US-05 Complete

+++ front-end/src/layout/tables/Table.js
@@ -1,43 +1,31 @@
-import React, { useState } from "react";
-import Modal from "../Modal";
+import React from "react";
 
 function Table({ table, deleteSeating }) {
-    const [modalIsOpen, setModalIsOpen] = useState(false);
-
     const status = table.reservation_id ? "Occupied" : "Free"
 
     let finishButton = null;
 
-    const closeModal = () => {
-        if (modalIsOpen) {
-            setModalIsOpen(false);
+    const displayConfirm = async () => {
+        document.getElementById(`finish`).blur();
+        if (window.confirm("Is this table ready to seat new guests? This cannot be undone.")) {
+            await deleteSeating(table);
         }
     }
 
-    const displayModal = () => {
-        setModalIsOpen(true);
-    }
-
-    const deleteAndClose = async () => {
-        await deleteSeating(table);
-        closeModal();
-    }
-
     if (status === "Occupied") {
-        finishButton = <button className="btn btn-danger" data-table-id-finish={table.table_id} onClick={displayModal}>Finish</button>;
+        finishButton = <button className="btn btn-danger" data-table-id-finish={table.table_id} id="finish" onClick={displayConfirm}>
+            Finish
+        </button>;
     }
 
     return (
-        <div className="card bg-dark text-white" onClick={closeModal}>
+        <div className="card bg-dark text-white">
             <div className="card-body">
                 <h3>{table.table_name}</h3>
                 <p>Capacity: {table.capacity}</p>
                 <p data-table-id-status={table.table_id}>{status}</p>
                 {finishButton}
             </div>
-            <Modal open={modalIsOpen} onClose={() => setModalIsOpen(false)} onOk={deleteAndClose}>
-                Is this table ready to seat new guests? This cannot be undone.
-            </Modal>
         </div>
     );
 }
